fixed encryption.py
