+++ lab_1/decryption.py
@@ -1,7 +1,7 @@
 import json
 import logging
 from collections import Counter
-from read_and_write import read_from_file, write_to_file
+
 
 logging.basicConfig(level=logging.INFO)
