+++ lab_1/encryption.py
@@ -1,8 +1,31 @@
 import json
 import logging
-from read_and_write import read_from_file, write_to_file
 
 logging.basicConfig(level=logging.INFO)
+
+
+def read_from_file(file_path: str) -> str:
+    """
+    Read text data from the specified file.
+    """
+    try:
+        with open(file_path, 'r', encoding='utf-8') as f:
+            return f.read()
+    except Exception as e:
+        logging.error(f"An error occurred while reading the file '{file_path}': {e}")
+        return ''
+
+
+def write_to_file(file_path: str, data: str) -> None:
+    """
+    Write data to the specified file.
+    """
+    try:
+        with open(file_path, 'w', encoding='utf-8') as f:
+            f.write(data)
+        logging.info(f"Data written to '{file_path}' successfully.")
+    except Exception as e:
+        logging.error(f"An error occurred while writing to the file '{file_path}': {e}")
 
 
 def encrypt_text(text_path: str, key_path: str) -> str:
@@ -10,7 +33,7 @@ def encrypt_text(text_path: str, key_path: str) -> str:
     Encrypts the given text using the provided key.
 
     Args:
-        text (str): The path to the text to be encrypted.
+        text_path (str): The path to the text to be encrypted.
         key_path (str): The path to the encryption key where keys are uppercase characters
                     and values are their corresponding encrypted forms.
 
@@ -19,8 +42,7 @@ def encrypt_text(text_path: str, key_path: str) -> str:
     """
     encrypted_text = ''
     try:
-        with open(text_path, 'r', encoding='utf-8') as f:
-            text = f.read()
+        text = read_from_file(text_path)
         with open(key_path, encoding='utf-8') as f:
             key = json.load(f)
         for char in text:
@@ -33,12 +55,12 @@ def encrypt_text(text_path: str, key_path: str) -> str:
     return encrypted_text
 
 
-def decrypt_text(encrypted_text: str, key_path: str) -> str:
+def decrypt_text(text_path: str, key_path: str) -> str:
     """
     Decrypts the given text using the provided key.
 
     Args:
-        encrypted_text (str): The path to the text to be decrypted.
+        text_path (str): The path to the text to be decrypted.
         key_path (str): The path to the encryption key where values are encrypted characters
                     and keys are their corresponding decrypted forms.
 
@@ -47,14 +69,13 @@ def decrypt_text(encrypted_text: str, key_path: str) -> str:
     """
     decrypted_text = ''
     try:
-        with open(encrypted_text, 'r', encoding='utf-8') as f:
-            text = f.read()
+        text = read_from_file(text_path)
         with open(key_path, encoding='utf-8') as f:
             key = json.load(f)
         reverse_key = {value: key for key, value in key.items()}
         for char in text:
-            if char in reverse_key:
-                decrypted_text += reverse_key[char]
+            if char.upper() in reverse_key:
+                decrypted_text += reverse_key[char.upper()]
             else:
                 decrypted_text += char
     except Exception as e:
@@ -67,14 +88,10 @@ if __name__ == "__main__":
         with open("lab_1/options.json", "r") as options_file:
             options = json.load(options_file)
         encrypted_text = encrypt_text(options['original_text'], options['key'])
+        write_to_file(options['encrypted_text'], encrypted_text)
 
-        with open(options['encrypted_text'], 'w', encoding='utf-8') as f:
-            f.write(encrypted_text)
-
-        decrypted_text = decrypt_text(
-            options['encrypted_text'], options['key'])
-
-        with open(options['decrypted_text'], 'w', encoding='utf-8') as f:
-            f.write(decrypted_text)
+        decrypted_text = decrypt_text(options['encrypted_text'], options['key'])
+        write_to_file(options['decrypted_text'], decrypted_text)
+        
     except Exception as e:
         logging.error(f"An error occurred: {e}")
